+++ avatar2/targets/openocd_target.py
@@ -69,7 +69,6 @@ class OpenOCDTarget(Target):
             self.log.debug("Resetting target...")
             openocd.reset()
 
-        self.protocols.set_all(openocd)
+        self.protocols.set_all(gdb)
         self.protocols.monitor = openocd
-        self.protocols.execution = gdb
         self.wait()